+++ korona/__pkg__.py
@@ -5,7 +5,7 @@
 __package_name__ = 'korona'
 __description__ = 'Package for building html pages'
 __url__ = 'https://github.com/bharadwajyarlagadda/korona'
-__version__ = '0.4.2'
+__version__ = '0.4.3.-dev'
 __author__ = 'Bharadwaj Yarlagadda'
 __email__ = 'yarlagaddabharadwaj@gmail.com'
 __license__ = 'MIT License'