+++ korona/html/tags/anchor.py
@@ -3,12 +3,95 @@
 
 from __future__ import absolute_import
 
-from ..root.attributes import TAG_ATTRIBUTES
-from ...lib.utils import validate_tag_attribute_value, validate_url
+from ...exceptions import TagAttributeError, AttributeValueError
+from ...warnings import warn
+from ...lib.utils import validate_url
 from ...templates.html.tags import anchor
 
 RECTANGLE_SHAPE_COORDINATES = 4
 CIRCLE_SHAPE_COORDINATES = 3
+CHARACTER_SETS = ['UTF-8', 'ISO-8859-1']
+
+ATTRIBUTES = {
+    'charset': {
+        'description': 'Specifies the character-set of a linked '
+                       'document.',
+        'values': None
+    },
+    'coords': {
+        'description': 'Specifies the coordinates of a link.',
+        'values': None
+    },
+    'download': {
+        'description': 'Specifies that the target will be downloaded '
+                       'when a user clicks on the hyper link.',
+        'values': None
+    },
+    'href': {
+        'descirption': 'Specifies the URL of the page the link goes '
+                       'to.',
+        'values': None
+    },
+    'hreflang': {
+        'description': 'Specifies the language of the linked '
+                       'document.',
+        'values': None
+    },
+    # TODO: Add media attribute.
+    'name': {
+        'description': 'Specifies the name of an anchor.',
+        'values': None
+    },
+    'rel': {
+        'description': 'Specifies the relationship between the current'
+                       ' document and the linked document.',
+        'values': ['alternate',
+                   'author',
+                   'bookmark',
+                   'help',
+                   'license',
+                   'next',
+                   'nofollow',
+                   'noreferrer',
+                   'prefetch',
+                   'prev',
+                   'search',
+                   'tag']
+    },
+    'rev': {
+        'description': 'Specifies the relationship between the linked '
+                       'document and the current document.',
+        'values': ['alternate',
+                   'stylesheet',
+                   'start',
+                   'next',
+                   'prev',
+                   'contents',
+                   'index',
+                   'glossary',
+                   'copyright',
+                   'chapter',
+                   'section',
+                   'subsection',
+                   'appendix',
+                   'help',
+                   'bookmark',
+                   'nofollow']
+    },
+    'shape': {
+        'description': 'Specifies the shape of a link.',
+        'values': ['default', 'rect', 'circle', 'poly']
+    },
+    'target': {
+        'description': 'Specifies where to open the linked document.',
+        'values': None
+    },
+    'type': {
+        'description': 'Specifies the media type of the linked '
+                       'document.',
+        'values': None
+    }
+}
 
 
 class A(object):
@@ -38,6 +121,12 @@ class A(object):
 
     .. versionchanged:: 0.3.1
         Added URL validation for href attribute.
+
+    .. versionchanged:: 0.4.3-dev
+        Now the user can be displayed with warnings for charset attribute.
+        Used custom exceptions for all the attributes. Changed the name of the
+        method from :func:`validate_values()` to
+        :func:`validate_attribute_values()`.
     """
     def __init__(self,
                  charset=None,
@@ -54,19 +143,21 @@ class A(object):
                  text=None):
         self.tag = 'a'
 
-        # Series of validation methods for validating all the attribute values
-        # given.
-        validate_tag_attribute_value(tag=self.tag, value=charset)
+        self.validate_charset(charset)
         coordinates = self.get_coords(shape=shape, coords=coords)
         self.pre_validate(href=href, attribute_name='download', value=download)
+        validate_url(attribute_name='download', url=download)
         validate_url(attribute_name='href', url=href)
         self.pre_validate(href=href, attribute_name='hreflang', value=hreflang)
-        self.validate_values(href=href, attribute_name='rel', value=rel)
-        self.validate_values(href=href, attribute_name='rev', value=rev)
-        self.validate_values(href=href, attribute_name='shape', value=shape)
+        self.pre_validate(href=href, attribute_name='rel', value=rel)
+        self.validate_attribute_values(attribute_name='rel',
+                                       attribute_value=rel)
+        self.validate_attribute_values(attribute_name='rev',
+                                       attribute_value=rev)
+        self.validate_attribute_values(attribute_name='shape',
+                                       attribute_value=shape)
         self.pre_validate(href=href, attribute_name='type', value=type)
 
-        # Assign all the values in a dictionary.
         self.values = {'charset': charset,
                        'coords': coordinates,
                        'download': download,
@@ -84,6 +175,22 @@ class A(object):
         """Returns the constructed anchor tag <a></a>."""
         return anchor.render(self.values)
 
+    def validate_charset(self, charset):
+        """Validates charset attribute. Warn the user showing the common
+        character sets used.
+        """
+        if not charset:
+            return
+
+        # If the attribute value is not a string raise a AttributeValueError.
+        if not isinstance(charset, str):
+            raise AttributeValueError('<{tag}>: {value} should be a string'
+                                      .format(tag=self.tag, value=charset))
+
+        if charset.upper() not in CHARACTER_SETS:
+            warn('Common character sets used are: {charsets}'
+                 .format(charsets=','.join(CHARACTER_SETS)))
+
     def get_coords(self, shape, coords):
         """Returns coordinates after a series of validations.
 
@@ -98,26 +205,26 @@ class A(object):
             return
 
         if (coords and not shape) or (shape and not coords):
-            raise AttributeError('<a>: shape attribute should be present when '
-                                 'coords are specified.')
+            raise TagAttributeError('<a>: shape attribute should be present '
+                                    'when coords are specified.')
 
         if isinstance(coords, str):
             coords = coords.split(',')
 
         if isinstance(coords, dict):
-            raise ValueError('<a>: {0} should be either list/tuple/str not a '
-                             'dictionary'
-                             .format(coords))
+            raise AttributeValueError('<a>: {0} should be either '
+                                      'list/tuple/str not a dictionary'
+                                      .format(coords))
 
         if shape == 'rect' and len(coords) != RECTANGLE_SHAPE_COORDINATES:
-            raise ValueError('<a>: {0} coordinates should be given for '
-                             'rectangle shape'
-                             .format(RECTANGLE_SHAPE_COORDINATES))
+            raise TagAttributeError('<a>: {0} coordinates should be given for '
+                                    'rectangle shape'
+                                    .format(RECTANGLE_SHAPE_COORDINATES))
 
         if shape == 'circle' and len(coords) != CIRCLE_SHAPE_COORDINATES:
-            raise ValueError('<a>: {0} coordinates should be given for '
-                             'circle shape'
-                             .format(CIRCLE_SHAPE_COORDINATES))
+            raise TagAttributeError('<a>: {0} coordinates should be given '
+                                    'for circle shape'
+                                    .format(CIRCLE_SHAPE_COORDINATES))
 
         return ','.join(str(coord) for coord in coords)
 
@@ -129,31 +236,27 @@ class A(object):
             return
 
         if not href:
-            raise AttributeError('<a>: {0} attribute is only used when '
-                                 'href attribute is set.'
-                                 .format(attribute_name))
+            raise TagAttributeError('<a>: {attribute_name} attribute is only '
+                                    'used when href attribute is set.'
+                                    .format(attribute_name=attribute_name))
 
-    def validate_values(self, href, attribute_name, value):
+    def validate_attribute_values(self, attribute_name, attribute_value):
         """Validates whether the given attribute value is a valid value or not.
         Some of the attributes have confined values. Even if we give some
         other value, the html output would not be correct.
         """
-        if not value:
+        if not attribute_value:
             return
 
-        if attribute_name == 'rel' and not href:
-            raise AttributeError('<a>: {attribute_name} is only used when '
-                                 'href attribute is set.'
-                                 .format(attribute_name=attribute_name))
+        if not isinstance(attribute_value, str):
+            raise AttributeValueError('<a>: {attribute_name} should be a '
+                                      'string value.'
+                                      .format(attribute_name=attribute_name))
 
-        if not isinstance(value, str):
-            raise ValueError('<a>: {0} should be a string value.'
-                             .format(attribute_name))
+        attribute_values = ATTRIBUTES[attribute_name]['values']
 
-        attribute_values = TAG_ATTRIBUTES[self.tag][attribute_name]['values']
-
-        if value not in attribute_values:
-            raise AttributeError('<a>: {attribute_name} attribute values '
-                                 'should be one of these: {values}'
-                                 .format(attribute_name=attribute_name,
-                                         values=','.join(attribute_values)))
+        if attribute_value not in attribute_values:
+            raise TagAttributeError('<a>: {attribute_name} attribute values '
+                                    'should be one of these: {values}'
+                                    .format(attribute_name=attribute_name,
+                                            values=','.join(attribute_values)))